Fixed Sobel calculations

+++ canny/labs/filters.h
@@ -6,7 +6,7 @@ void Conv2DSerial(float *inImg, float *outImg, double filter[FILTERSIZE][FILTERS
 void populate_blur_filter(double outFilter[FILTERSIZE][FILTERSIZE]);
 void GradientSobelSerial(float *inImg, float *sobelImg, float *gradientImg, int height, int width); 
 #ifdef __CUDACC__
-__global__ void ColorToGrayscale(float *inImg, int *outImg, int width, int height); 
+__global__ void ColorToGrayscale(float *inImg, float *outImg, int width, int height); 
 #endif
 
 #endif // __FILTERS_H__
